Created a secondary twitter strategy.
This allows you to have two different Twitter apps associated with a user of your application. The use case I encountered was:
    
As a user
I want to use a Twitter account to login to the app
And I want to use a different Twitter account from which to post updates.